# docs/swagger.yaml
basePath: /api
definitions:
  api_err.ApiErr:
    properties:
      msg:
        type: string
      typ:
        $ref: '#/definitions/api_err.ApiErrTyp'
    type: object
  api_err.ApiErrTyp:
    enum:
    - not_found
    - bad_input
    - forbidden
    - unauthorized
    - already_claimed
    - pending_claim_exists
    - username_invalid
    - username_taken
    - report_target_missing
    - report_locked
    - duplicate_open_report
    - text_too_long
    - unknown
    type: string
    x-enum-varnames:
    - AetNotFound
    - AetBadInput
    - AetForbidden
    - AetUnauthorized
    - AetAlreadyClaimed
    - AetPendingClaimExists
    - AetUsernameInvalid
    - AetUsernameTaken
    - AetReportTargetMissing
    - AetReportLocked
    - AetDuplicateOpenReport
    - AetTextTooLong
    - AetUnknown
...
  models.AdminMini:
    properties:
      avatar_key:
        type: string
      id:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
...
  models.PublicUser:
    properties:
      avatar_key:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  models.ReportStatus:
    enum:
    - pending
    - in_progress
    - dismissed
    - actioned
    type: string
    x-enum-varnames:
    - ReportStatusPending
    - ReportStatusInProgress
    - ReportStatusDismissed
    - ReportStatusActioned
  models.ReportTargetType:
    enum:
    - review
    - place
    type: string
    x-enum-varnames:
    - ReportTargetReview
    - ReportTargetPlace
  models.ReportType:
    enum:
    - spam
    - misleading
    - inappropriate
    - profanity
    type: string
    x-enum-varnames:
    - ReportTypeSpam
    - ReportTypeMisleading
    - ReportTypeInappropriate
    - ReportTypeProfanity
...
  models.UserMini:
    properties:
      avatar_key:
        type: string
      id:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
...
  services.CreateReviewInput:
    properties:
      images:
        items:
          type: string
        type: array
      price_rating:
        type: integer
      quality_rating:
        type: integer
      star_rating:
        type: integer
      text:
        type: string
    type: object
...
  services.EditReportInput:
    properties:
      text:
        type: string
      type:
        $ref: '#/definitions/models.ReportType'
    type: object
  services.MeView:
    properties:
      avatar_key:
        type: string
      blocked:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owns_place:
        type: boolean
      phone:
        type: string
      username:
        type: string
    type: object
...
  services.PublicUserView:
    properties:
      review_count:
        type: integer
      user:
        $ref: '#/definitions/models.PublicUser'
    type: object
...
  services.ReportTarget:
    properties:
      avatar_key:
        type: string
      content:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/models.ReportTargetType'
    type: object
...
  services.ReportView:
    properties:
      admin:
        $ref: '#/definitions/models.AdminMini'
      admin_id:
        type: string
      admin_response:
        type: string
      created_at:
        type: string
      id:
        type: string
      reported_user:
        $ref: '#/definitions/models.UserMini'
      reported_user_id:
        type: string
      reporter_user:
        $ref: '#/definitions/models.UserMini'
      reviewed_at:
        type: string
      status:
        $ref: '#/definitions/models.ReportStatus'
      target:
        $ref: '#/definitions/services.ReportTarget'
      target_id:
        type: string
      target_type:
        $ref: '#/definitions/models.ReportTargetType'
      text:
        type: string
      type:
        $ref: '#/definitions/models.ReportType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  services.ReviewReportInput:
    properties:
      admin_response:
        type: string
      block_reported_user:
        type: boolean
      delete_target_review:
        type: boolean
      status:
        $ref: '#/definitions/models.ReportStatus'
    type: object
  services.ReviewView:
    properties:
      created_at:
        type: string
      id:
        type: string
      images:
        items:
          type: string
        type: array
      latest:
        type: boolean
      place_id:
        type: string
      price_rating:
        type: integer
      quality_rating:
        type: integer
      star_rating:
        type: integer
      text:
        type: string
      user:
        $ref: '#/definitions/models.UserMini'
      user_id:
        type: string
    type: object
...
  services.SubmitReportInput:
    properties:
      target_id:
        type: string
      target_type:
        $ref: '#/definitions/models.ReportTargetType'
      text:
        type: string
      type:
        $ref: '#/definitions/models.ReportType'
    type: object
  services.UpdateMeInput:
    properties:
      avatar_key:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: S101 backend API
  title: S101 API
  version: "1.0"
paths:
...
  /admin/reports:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      - description: Filter by target type
        in: query
        name: target_type
        type: string
      - description: Filter by target ID
        in: query
        name: target_id
        type: string
      - description: Filter by reported user
        in: query
        name: reported_user_id
        type: string
      - description: Filter by admin
        in: query
        name: admin_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: list reports'
      tags:
      - admin-reports
  /admin/reports/{id}:
    get:
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.ReportView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: get report by ID'
      tags:
      - admin-reports
    put:
      consumes:
      - application/json
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Review decision
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.ReviewReportInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.ReportView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: review a report'
      tags:
      - admin-reports
...
  /places/{id}/reviews:
    get:
      parameters:
      - description: Place UUID
        in: path
        name: id
        type: string
      - description: 'Return full history (default: latest only)'
        in: query
        name: all
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      summary: List reviews for a place
      tags:
      - reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Place UUID
        in: path
        name: id
        required: true
        type: string
      - description: Review data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CreateReviewInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.ReviewView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Create a review for a place
      tags:
      - reviews
...
  /reports:
    post:
      consumes:
      - application/json
      parameters:
      - description: Report data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.SubmitReportInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.ReportView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Submit a report
      tags:
      - reports
  /reports/{id}:
    delete:
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Delete own report
      tags:
      - reports
    get:
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.ReportView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Get own report by ID
      tags:
      - reports
    put:
      consumes:
      - application/json
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.EditReportInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.ReportView'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Edit own report
      tags:
      - reports
...
  /reports/mine:
    get:
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: List own reports
      tags:
      - reports
  /reviews/{id}:
    delete:
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Delete own review
      tags:
      - reviews
  /users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.PublicUserView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      summary: Get public user profile
      tags:
      - users
  /users/{id}/reviews:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      summary: List reviews by a user
      tags:
      - users
  /users/me:
    delete:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Delete own account
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.UpdateMeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.MeView'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Update own profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"